Remove mkboot entirely
mkboot has become increasngly distant from accepted practices on
modern Debian isntallations, and has limited utility. mkboot only
works on i386 and ia64.  If GRUB is installed, it does nothing. It is
unsure how the grub2 detection works. mkboot only works on floppy
diskette drives, which are increasingly rare beasts.  If rdev is not
present, mkboot will fail to determine the correct root partition.

In most cases, running mkboot is the wrong thing to do, and this
bugginess, in conjunction with its irrelevancy, require that this be
removed.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>