Handle symlinks in /boot when reinstalling
The check to see if we need to handle symbolic links in boot in
installkernel, meant to check whether the current link points to a
different file than the one being installed, works fine provided that we
are indeed not reinstalling the same kernel version. Otherwise, the
inequality test will fail and we'll fall back to the else clause of the
if statement, leading to plain file creation again.

Nested the if statements.

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>