Ignore a fifth argument to installkernel
While parsing the command line options, we check to see if the number of
options is sane. On most architectures, this is either three or four
options; except of course, powerpc has to be all different, and passes in
a fifth argument, because it is "special". We now ignore the fifth
argument, and do not flag is as an error, which it would be for any arch
apart from powerpc

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>