Permission and snity checks run on link targets, not symlinks
run-parts should ignore symbolic links to directories, since the
behaviour of run-parts in this scenario is hard to define. Indeed, just
like permission checks, all sanity checks should run on the link targtet,
not on the link, since the link target is what ends up getting executed.

The exception is the name checks; this allows people to set a symbolic
link to a target that would otherwise not pass teh run-parts file naming
convention. Thus, the valid name check is performed on the link name, and
all other checks are performed on the link target.

Signed-Off-By: Manoj Srivastava <srivasta@debian.org>
Signed-off-by: Manoj Srivastava <srivasta@debian.org>